Experimenting with abstract
