+++ src/main/kotlin/gg/flyte/twilight/boards/scoreboards/DynamicScoreboard.kt
@@ -1,0 +1,2 @@
+package gg.flyte.twilight.boards.scoreboards
+
